chore: combine santisation with reference link
