+++ packages/components/src/templates/next/components/complex/Image/Image.tsx
@@ -2,11 +2,7 @@ import type { PropsWithChildren } from "react"
 
 import type { ImageProps } from "~/interfaces"
 import { tv } from "~/lib/tv"
-import {
-  getReferenceLinkHref,
-  getSanitizedLinkHref,
-  isExternalUrl,
-} from "~/utils"
+import { getReferenceLinkHref, isExternalUrl } from "~/utils"
 import { Link } from "../../internal/Link"
 import { ImageClient } from "./ImageClient"
 
@@ -49,7 +45,7 @@ const ImageContainer = ({
   <div className={compoundStyles.container()}>
     {href !== undefined ? (
       <Link
-        href={getSanitizedLinkHref(href)}
+        href={href}
         target={isExternalUrl(href) ? "_blank" : undefined}
         rel={isExternalUrl(href) ? "noopener nofollow" : undefined}
         LinkComponent={LinkComponent}
