+++ packages/components/src/templates/next/components/complex/InfoCards/InfoCards.tsx
@@ -11,7 +11,6 @@ import type {
 import { tv } from "~/lib/tv"
 import {
   getReferenceLinkHref,
-  getSanitizedLinkHref,
   getTailwindVariantLayout,
   isExternalUrl,
 } from "~/utils"
@@ -103,7 +102,7 @@ const InfoCardContainer = ({
 >): JSX.Element => {
   return url ? (
     <Link
-      href={getSanitizedLinkHref(getReferenceLinkHref(url, site.siteMap))}
+      href={getReferenceLinkHref(url, site.siteMap)}
       className={compoundStyles.cardContainer()}
       LinkComponent={LinkComponent}
     >
