+++ packages/components/src/templates/next/components/complex/InfoCols/InfoCols.tsx
@@ -6,11 +6,7 @@ import type { SupportedIconName } from "~/common/icons"
 import type { InfoColsProps } from "~/interfaces"
 import { SUPPORTED_ICONS_MAP } from "~/common/icons"
 import { tv } from "~/lib/tv"
-import {
-  getReferenceLinkHref,
-  getSanitizedLinkHref,
-  getTailwindVariantLayout,
-} from "~/utils"
+import { getReferenceLinkHref, getTailwindVariantLayout } from "~/utils"
 import { groupFocusVisibleHighlightNonRac } from "~/utils/rac"
 import { ComponentContent } from "../../internal/customCssClass"
 import { Link } from "../../internal/Link"
@@ -78,9 +74,7 @@ const InfoBoxes = ({
         ({ title, icon, description, buttonUrl, buttonLabel }, idx) => (
           <Link
             LinkComponent={LinkComponent}
-            href={getSanitizedLinkHref(
-              getReferenceLinkHref(buttonUrl, site.siteMap),
-            )}
+            href={getReferenceLinkHref(buttonUrl, site.siteMap)}
             key={idx}
             className={compoundStyles.infoBox()}
           >
