+++ packages/components/src/utils/index.ts
@@ -4,7 +4,6 @@ export { getFormattedDate } from "./getFormattedDate"
 export { getRandomNumberBetIntervals } from "./getRandomNumber"
 export { getReferenceLinkHref } from "./getReferenceLinkHref"
 export { getSanitizedIframeWithTitle } from "./getSanitizedIframeWithTitle"
-export { getSanitizedLinkHref } from "./getSanitizedLinkHref"
 export { getSiderailFromSiteMap } from "./getSiderailFromSiteMap"
 export { getSitemapAsArray } from "./getSitemapAsArray"
 export { getTailwindVariantLayout } from "./getTailwindVariantLayout"
